REFACTOR : change in destroy(EnemyShip) and hit(), and other changes
1.
void destroy(EnemyShip) ---> void loseHp_or_destroy(EnemyShip)
in EnemyShipFormation.java

Since previous function (destroy()) has a role to check if the EnemyShip should lose its hp or to die, changed this function's name.

This name should be changed because it is too long and non-intuitive.
It would be easy since it only used in one line; line #290 in GameScreen.java

2.
boolean hit() ---> void hit()
in EnemyShip.java

isDestroyed() has same role to check if the ship is destroyed.
So, I seperated functions' role. Now hit() only makes changes to ship's hp.

3. how it works - loseHp_or_destroy()
Function checks the list of EnemyShips.
If it finds the ship (to lose hp or be destroyed), hit() activates and checks it is destroyed.
If it is, prints log messege.
Otherwise, it means it didn't got destroyed. So it means it only lost Hp.

4.
Fix : bug prints too many "hp lost" log messeges.

5.
Getter and Setter for Hp is no more used but not deleted.