FIX: fixed game freezing bug after destroyed all ships
There is a bug that the game stops after all enemys are destroyed. Here's reason why:

if ((this.enemyShipFormation.isEmpty() || this.lives == 0) && !this.levelFinished) {
			this.levelFinished = true;
			this.screenFinishedCooldown.reset();
		}

In this code, the game ends when there are no more enemyShips.
The number of enemyShips are tracked with variable 'shipCount' in 'entity/EnemyshipFormation.java'.
But 'EnemyshipFormation._destroy(Bullet, EnemyShip, boolean)' only reduced 'shipCount' when ship's Hp is 0 or below.
So eventhough all enemys were destroyed, the game just stopped.
So, add setHp(0) in 'chainExplosion()' to fix this.